Add files via upload
The water in daily life is an imperative and needful resource to human beings and all other living organisms. The contamination of water by adding pollutants leads to noxious effects and also it is the new way of introducing disease to the people. Thus machine learning plays an important role in Predicting the water whether it is in good state or in bad condition. The machine learning algorithm is used to predict the water quality and also statistical accuracy parameters are calculated for each machine learning algorithm. SVM, KNN, Decision Tree classifiers, XGBoost, Ada boost are used to predict the condition of water. The XG Boost algorithm predicts the water quality with 97% accuracy .By evaluating the values of accuracy, Precision values and more statistical parameters are measured and compared.